python/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# GitHub API
PyGithub>=2.1.1

# OpenAI
openai>=1.12.0

# Embeddings
sentence-transformers>=2.3.1
torch>=2.1.0

# Vector Database
pinecone>=5.0.0

# Environment & Config
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# File Handling
pathspec>=0.12.1

# HTTP Client
httpx>=0.26.0
